refactored front end routes into frontend folder and frontend javascript input handlers
